Windows deployment works pretty well
