+++ scrutiny/__main__.py
@@ -13,7 +13,24 @@
 # nuitka-project-else:
 #     nuitka-project: --windows-console-mode=disabled
 
-from scrutiny.entry_points.scrutiny import scrutiny_cli
+from scrutiny.cli import CLI
+import sys
+import os
 
+def scrutiny_server() -> None:
+    cli = CLI(os.getcwd())
+    code = cli.run(['server'] + sys.argv[1:])
+    sys.exit(code)
+
+def scrutiny_gui_with_server() -> None:
+    cli = CLI(os.getcwd())
+    code = cli.run(['gui', '--start-local-server', '--auto-connect'] + sys.argv[1:])
+    sys.exit(code)
+
+def scrutiny_cli()-> None:
+    cli = CLI(os.getcwd())
+    code = cli.run(sys.argv[1:])
+    sys.exit(code)
+    
 if __name__ == '__main__':
     scrutiny_cli()
